# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 61bd7d167a739847d5cfd096
  title: schoolmng
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/course/{id}":
    delete:
      x-skaffolder-id: 61bd80927a739847d5cfd163
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61bd7ee77a739847d5cfd12f
      x-skaffolder-resource: Course
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61bd80927a739847d5cfd158
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61bd7ee77a739847d5cfd12f
      x-skaffolder-resource: Course
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    post:
      x-skaffolder-id: 61bd80927a739847d5cfd15a
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61bd7ee77a739847d5cfd12f
      x-skaffolder-resource: Course
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
  "/exam/{id}/Validate":
    post:
      x-skaffolder-id: 61bd81977a739847d5cfd16f
      x-skaffolder-name: Validate
      x-skaffolder-id-resource: 61bd7f247a739847d5cfd136
      x-skaffolder-resource: Exam
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: this is used to validate the exam
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Boolean
      x-skaffolder-url: /{id}/Validate
      x-skaffolder-roles:
      summary: this is used to validate the exam
      tags:
        - Exam
      parameters:
        - name: id
          x-skaffolder-type: String
          in: path
          description: ID of the exam
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  "/exam/{id}":
    delete:
      x-skaffolder-id: 61bd80b07a739847d5cfd16d
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61bd7f247a739847d5cfd136
      x-skaffolder-resource: Exam
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61bd80b07a739847d5cfd168
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61bd7f247a739847d5cfd136
      x-skaffolder-resource: Exam
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    post:
      x-skaffolder-id: 61bd80b07a739847d5cfd166
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61bd7f247a739847d5cfd136
      x-skaffolder-resource: Exam
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Exam"
  "/exam/findBy_Course/{key}":
    get:
      x-skaffolder-id: 61bd80927a739847d5cfd15c
      x-skaffolder-name: findBy_Course
      x-skaffolder-id-resource: 61bd7f247a739847d5cfd136
      x-skaffolder-resource: Exam
      x-skaffolder-crudAction: findBy_Course
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_Course
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_Course/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_Course
      tags:
        - Exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
  "/exam/findBy_Student/{key}":
    get:
      x-skaffolder-id: 61bd80657a739847d5cfd143
      x-skaffolder-name: findBy_Student
      x-skaffolder-id-resource: 61bd7f247a739847d5cfd136
      x-skaffolder-resource: Exam
      x-skaffolder-crudAction: findBy_Student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_Student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_Student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_Student
      tags:
        - Exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
  "/exam/findBy_Teacher/{key}":
    get:
      x-skaffolder-id: 61bd807e7a739847d5cfd150
      x-skaffolder-name: findBy_Teacher
      x-skaffolder-id-resource: 61bd7f247a739847d5cfd136
      x-skaffolder-resource: Exam
      x-skaffolder-crudAction: findBy_Teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_Teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_Teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_Teacher
      tags:
        - Exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Course:
      x-skaffolder-id: 61bd7ee77a739847d5cfd12f
      x-skaffolder-id-db: 61bd7d167a739847d5cfd097
      x-skaffolder-id-entity: 61bd7ee77a739847d5cfd12d
      x-skaffolder-url: /course
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 61bd7ee77a739847d5cfd12e
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - name
    Exam:
      x-skaffolder-id: 61bd7f247a739847d5cfd136
      x-skaffolder-id-db: 61bd7d167a739847d5cfd097
      x-skaffolder-id-entity: 61bd7f247a739847d5cfd133
      x-skaffolder-url: /exam
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Score:
          type: number
          x-skaffolder-id-attr: 61bd7f247a739847d5cfd134
          x-skaffolder-type: Number
          x-skaffolder-required: true
        place:
          type: string
          x-skaffolder-id-attr: 61bd7f247a739847d5cfd135
          x-skaffolder-type: String
          x-skaffolder-required: true
        valid:
          type: boolean
          x-skaffolder-id-attr: 61bdd4e27a739847d5cfd181
          x-skaffolder-type: Boolean
      x-skaffolder-relations:
        _Course:
          x-skaffolder-id: 61bd7fc37a739847d5cfd139
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61bd7f247a739847d5cfd133
          x-skaffolder-ent2: 61bd7ee77a739847d5cfd12d
        _Student:
          x-skaffolder-id: 61bd7fc37a739847d5cfd138
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61bd7f247a739847d5cfd133
          x-skaffolder-ent2: 61bd7e7d7a739847d5cfd120
        _Teacher:
          x-skaffolder-id: 61bd7fc37a739847d5cfd13a
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61bd7f247a739847d5cfd133
          x-skaffolder-ent2: 61bd7ebd7a739847d5cfd127
      required:
        - _id
        - Score
        - place
...